refactor(mobile): Extract TemplateGrid into separate component
- Created reusable TemplateGrid component
- Moved template rendering logic to dedicated component
- Improved code organization and reusability
- Reduced profile screen complexity
